Hotfix an oversight from my leash proc that will cause runtimes
When optimizing the earlier checks that see if they're still nearby and that they belong to the person, I neglected the fact that, after that block, we still don't know for sure if the leash even has an assigned pet. This will runtime for people holding a mix of connected and unconnected leashes.